feat: Update Sample data model and related logic
This commit refactors the `Sample` data model and updates its usage across the application, including repositories, ViewModels, and tests.

- Changed the `id` field in the `Sample` data class from `Int` to `String` for better compatibility with Firestore document IDs.
- Added new fields to the `Sample` model:
    - `usersLike: List<String>`
    - `isPublic: Boolean`
    - `storageZipPath: String`
    - `storageImagePath: String`
    - `storagePreviewSamplePath: String`
- Updated `SampleRepositoryFirebase` to handle the new `String` ID and additional fields for both data mapping and retrieval.
- Implemented `getSample(sampleId: String)` in the sample repositories to fetch a single sample by its ID.
- Adjusted all test cases and mock data initializations to align with the updated `Sample` model structure.

+++ app/src/main/java/com/neptune/neptune/ui/search/SearchScreen.kt
@@ -126,7 +126,8 @@ fun ScrollableColumnOfSamples(
         val width = 300.dp
         items(samples) { sample ->
           // change height and width if necessary
-          val testTags = SearchScreenTestTagsPerSampleCard(idInColumn = sample.id)
+          val testTags =
+              SearchScreenTestTagsPerSampleCard(idInColumn = sample.id.toInt()) // TODO CHANGE THIS
           SampleCard(
               sample = sample,
               width = width,
